requirements.txt
# Core
flask>=3.0
requests>=2.31

# LLM + embeddings
sentence-transformers>=2.7
rank-bm25>=0.2

# PDF parsing
PyMuPDF>=1.24

# MCP server/client
fastmcp>=0.2

# Optional config formats (for --pdf_config .yaml/.yml)
pyyaml>=6.0

# Optional speed-up for vector search (if wheels are available for your OS/Python)
# faiss-cpu

# === Core dependencies ===
pandas>=2.0.0

# === Model backend ===
llama-cpp-python>=0.2.90   # for Llama model loading and inference

# === CSV / data utilities ===
numpy>=1.26.0

# === Optional (for local performance / CPU threading) ===
tqdm>=4.65.0                # progress bars (optional but helpful)

# === Logging & system handling ===
psutil>=5.9.0
